Move action and keyboard helpers to mixin classes (#307)
